🐛 don't die on c==-1 or bad date; isoformat dates; output valid json
--json flag to attempt to put valid json on stdout
as a result, the script can work without either jmespath or json (but not both)

print to stderr for warnings in notifications so we can pipe to jq
ignore lines that are file ids (instead of crash with failed assert)

mozidb.py:_decode_string() doesn't handle c == -1. just skips the whole thing
mozserial.py:start_read() previously failed on bad date
                          now just use start of unix epoch (timestamp=0)

editor ate some of the tabs. unnecessary diffs as a result